DEV: moved K2700 to keithely.py and added K6514
